Don't use `Aggregate` as the aggregate class
This feels awkward no matter what option we choose, but this feels
the least awkward. It doesn't make a whole lot of sense for the
repository to call `Aggregate.new` and now we instead choose to
call it by its semantic name, e.g. `Appointment` or `Person`. It
does mean the namespacing would be something like
`Appointment::Appointment`, which feels weird. It feels less weird
than having only `Appointment` as the aggregate, since that would
mean we'd have to nest the respository and commands below the
aggregate, which clearly is not ideal.