Add a Unicorn configuration file
There's the potential for this to differ from staging to production and
across various development environments (for instance devs using Boxen)
so let's add it to the gitignore file so it doesn't get stored in
source control. Instead let's add an example file to SCM so peeps can
get started quickly.